Enable tweaking the size of the browser window
window_size is the size of the QWebView widget containing the QWebPage
object.  The problem was that when "viewport=full" was selected,
viewport was not resized till the very rendering and inherited the
default window size which is small (640x480).  In presence of so called
fluid layouts this could significantly affect the appearance of the site
and there was no way to tweak that initial size, because "viewport="
argument was already occupied with "viewport=full".

* add "window_size=" query argument to "render.xyz" endpoints

* add "splash:set_window_size" function to Lua environment

* increase default window size from 640x480 to 1024x768

* if "splash:set_viewport('full')" fails in Lua environment, raise an
  error instead of silently falling back to default size