More work on PR: remove splash global variable.
I tried to remove the splash global variable _before_ calling the user's
code (and stashing it in a variable in the closure scope) but I don't
think it is possible with QT4: it leads to segfaults.  I think QT notices
when we delete the global variable and it destroys the underlying C++
object. In QT5 I think we could work around it by specifying a different
ownership policy, but for now I've rewritten the wrapper so that the GV
is deleted after the user's code calls us back.