First round of unit tests for splash:wait_for_resume.
Added unit tests for the existing behavior of splash:wait_for_resume
and a couple of expected fail (xfail) tests for handling different
types of arguments to splash.resume(). I also updated the docs to
reflect a few fixes and changing the argument 'msg' to 'value' in
anticipation of being able to return different types.

Added some separate tests for OneShotCallbackProxy: it's difficult
(or impossible) to test some of the behavior through the HTTP API,
so this small test suite invokes a callback proxy and tests some
of its behavior outside of any HTTP request context.