Allow resume() to accept different types of argument.
resume() was previously defined with a 'str' argument that results in
QT stringifying any value or object passed to it. In this commit, I've
added support for all of the basic Lua types to be passed from JS to
resume() and then on to Lua: nil, number, boolean, string, and table,
including unit tests for each type.

There are some tradeoffs with this approach. See my extended comments
in the PR thread.