Remove my custom QT type conversion and call qt2py instead.
Mikhail and I agreed that we will not to try to support returning an Array
to resume(), since that requires some voodoo. Instead, I added documentation
showing the surprising results in the array case.

These two changes make for a nice cleanup. I cannot replicate the Travis
build failures on my local machine, but the failures were all inside the QT
type conversion code anyway, so I expect that this commit will result
in a successful build.