[tests/test_execute] dont rely on webkit version quirks
some tests were failing in my machine because of minor differences in error messages
that probably depend on webkit version or OS. For example js tests were
failing for me because syntax error message was "Syntax Error:
unexpected token" and not "Syntax Error: parse error". Or in set_content
I got "about:blank" in url if request to splash was without url
argument, and test expected url field to be empty.