[browser_tab] delete NAT after using it
this ensures there is no segfault in case of race conditions. It looks like requests can be downloaded
by NAT even after render is done and finished. If deleteLater() is not called it seems like render is done, but NAT
is still in memory and is still in use, when executing callbacks it probably tries to access something
that was deleted and it ends SEGFAULT